Add `BaseDataset` as parent.
1. Added BaseDataset as parent for `Dataset` and `HeteroGraphDataset`
   class and moved all the basic functionaluty like process, download,
   and name utilities to parent class
2. Removed all instances of `HeteroGraphData` from dataset
3. Cahnged data type of node feature in DBLP and IMDB dataset.
